+++ mplapack/reference/Rptrfs.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2008-2021
+ * Copyright (c) 2008-2025
  *      Nakata, Maho
  *      All rights reserved.
  *
@@ -26,6 +26,13 @@
  *
  */
 
+// Derived from LAPACK routine DPTRFS.
+// Original LAPACK authors:
+//   Univ. of Tennessee
+//   Univ. of California Berkeley
+//   Univ. of Colorado Denver
+//   NAG Ltd.
+
 #include <mpblas.h>
 #include <mplapack.h>
 
@@ -38,7 +45,7 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
     REAL safe1 = 0.0;
     REAL safe2 = 0.0;
     INTEGER count = 0;
-    const REAL three = 3.0e+0;
+    const REAL three = 3.0;
     REAL lstres = 0.0;
     REAL bi = 0.0;
     REAL dx = 0.0;
@@ -46,35 +53,12 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
     INTEGER i = 0;
     REAL cx = 0.0;
     REAL s = 0.0;
-    const REAL two = 2.0e+0;
+    const REAL two = 2.0;
     const INTEGER itmax = 5;
     const REAL one = 1.0;
     INTEGER ix = 0;
     //
-    //  -- LAPACK computational routine --
-    //  -- LAPACK is a software package provided by Univ. of Tennessee,    --
-    //  -- Univ. of California Berkeley, Univ. of Colorado Denver and NAG Ltd..--
-    //
-    //     .. Scalar Arguments ..
-    //     ..
-    //     .. Array Arguments ..
-    //     ..
-    //
-    //  =====================================================================
-    //
-    //     .. Parameters ..
-    //     ..
-    //     .. Local Scalars ..
-    //     ..
-    //     .. External Subroutines ..
-    //     ..
-    //     .. Intrinsic Functions ..
-    //     ..
-    //     .. External Functions ..
-    //     ..
-    //     .. Executable Statements ..
-    //
-    //     Test the input parameters.
+    // Test the input parameters.
     //
     info = 0;
     if (n < 0) {
@@ -91,7 +75,7 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
         return;
     }
     //
-    //     Quick return if possible
+    // Quick return if possible
     //
     if (n == 0 || nrhs == 0) {
         for (j = 1; j <= nrhs; j = j + 1) {
@@ -101,7 +85,7 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
         return;
     }
     //
-    //     NZ = maximum number of nonzero elements in each row of A, plus 1
+    // NZ = maximum number of nonzero elements in each row of A, plus 1
     //
     nz = 4;
     eps = Rlamch("Epsilon");
@@ -109,7 +93,7 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
     safe1 = nz * safmin;
     safe2 = safe1 / eps;
     //
-    //     Do for each right hand side
+    // Do for each right hand side
     //
     for (j = 1; j <= nrhs; j = j + 1) {
         //
@@ -117,10 +101,10 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
         lstres = three;
     statement_20:
         //
-        //        Loop until stopping criterion is satisfied.
+        // Loop until stopping criterion is satisfied.
         //
-        //        Compute residual R = B - A * X.  Also compute
-        //        abs(A)*abs(x) + abs(b) for use in the backward error bound.
+        // Compute residual R = B - A * X.  Also compute
+        // abs(A)*abs(x) + abs(b) for use in the backward error bound.
         //
         if (n == 1) {
             bi = b[(j - 1) * ldb];
@@ -148,34 +132,34 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
             work[n - 1] = abs(bi) + abs(cx) + abs(dx);
         }
         //
-        //        Compute componentwise relative backward error from formula
+        // Compute componentwise relative backward error from formula
         //
-        //        max(i) ( abs(R(i)) / ( abs(A)*abs(X) + abs(B) )(i) )
+        // max(i) ( abs(R(i)) / ( abs(A)*abs(X) + abs(B) )(i) )
         //
-        //        where abs(Z) is the componentwise absolute value of the matrix
-        //        or vector Z.  If the i-th component of the denominator is less
-        //        than SAFE2, then SAFE1 is added to the i-th components of the
-        //        numerator and denominator before dividing.
+        // where abs(Z) is the componentwise absolute value of the matrix
+        // or vector Z.  If the i-th component of the denominator is less
+        // than SAFE2, then SAFE1 is added to the i-th components of the
+        // numerator and denominator before dividing.
         //
         s = zero;
         for (i = 1; i <= n; i = i + 1) {
             if (work[i - 1] > safe2) {
-                s = max(s, REAL(abs(work[(n + i) - 1]) / work[i - 1]));
+                s = max(s, abs(work[(n + i) - 1]) / work[i - 1]);
             } else {
-                s = max(s, REAL((abs(work[(n + i) - 1]) + safe1) / (work[i - 1] + safe1)));
+                s = max(s, (abs(work[(n + i) - 1]) + safe1) / (work[i - 1] + safe1));
             }
         }
         berr[j - 1] = s;
         //
-        //        Test stopping criterion. Continue iterating if
-        //           1) The residual BERR(J) is larger than machine epsilon, and
-        //           2) BERR(J) decreased by at least a factor of 2 during the
-        //              last iteration, and
-        //           3) At most ITMAX iterations tried.
+        // Test stopping criterion. Continue iterating if
+        // 1) The residual BERR(J) is larger than machine epsilon, and
+        // 2) BERR(J) decreased by at least a factor of 2 during the
+        // last iteration, and
+        // 3) At most ITMAX iterations tried.
         //
         if (berr[j - 1] > eps && two * berr[j - 1] <= lstres && count <= itmax) {
             //
-            //           Update solution and try again.
+            // Update solution and try again.
             //
             Rpttrs(n, 1, df, ef, &work[(n + 1) - 1], n, info);
             Raxpy(n, one, &work[(n + 1) - 1], 1, &x[(j - 1) * ldx], 1);
@@ -184,23 +168,23 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
             goto statement_20;
         }
         //
-        //        Bound error from formula
+        // Bound error from formula
         //
-        //        norm(X - XTRUE) / norm(X) .le. FERR =
-        //        norm( abs(inv(A))*
-        //           ( abs(R) + NZ*EPS*( abs(A)*abs(X)+abs(B) ))) / norm(X)
+        // norm(X - XTRUE) / norm(X) .le. FERR =
+        // norm( abs(inv(A))*
+        // ( abs(R) + NZ*EPS*( abs(A)*abs(X)+abs(B) ))) / norm(X)
         //
-        //        where
-        //          norm(Z) is the magnitude of the largest component of Z
-        //          inv(A) is the inverse of A
-        //          abs(Z) is the componentwise absolute value of the matrix or
-        //             vector Z
-        //          NZ is the maximum number of nonzeros in any row of A, plus 1
-        //          EPS is machine epsilon
+        // where
+        // norm(Z) is the magnitude of the largest component of Z
+        // inv(A) is the inverse of A
+        // abs(Z) is the componentwise absolute value of the matrix or
+        // vector Z
+        // NZ is the maximum number of nonzeros in any row of A, plus 1
+        // EPS is machine epsilon
         //
-        //        The i-th component of abs(R)+NZ*EPS*(abs(A)*abs(X)+abs(B))
-        //        is incremented by SAFE1 if the i-th component of
-        //        abs(A)*abs(X) + abs(B) is less than SAFE2.
+        // The i-th component of abs(R)+NZ*EPS*(abs(A)*abs(X)+abs(B))
+        // is incremented by SAFE1 if the i-th component of
+        // abs(A)*abs(X) + abs(B) is less than SAFE2.
         //
         for (i = 1; i <= n; i = i + 1) {
             if (work[i - 1] > safe2) {
@@ -212,39 +196,39 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
         ix = iRamax(n, work, 1);
         ferr[j - 1] = work[ix - 1];
         //
-        //        Estimate the norm of inv(A).
+        // Estimate the norm of inv(A).
         //
-        //        Solve M(A) * x = e, where M(A) = (m(i,j)) is given by
+        // Solve M(A) * x = e, where M(A) = (m(i,j)) is given by
         //
-        //           m(i,j) =  abs(A(i,j)), i = j,
-        //           m(i,j) = -abs(A(i,j)), i .ne. j,
+        // m(i,j) =  abs(A(i,j)), i = j,
+        // m(i,j) = -abs(A(i,j)), i .ne. j,
         //
-        //        and e = [ 1, 1, ..., 1 ]**T.  Note M(A) = M(L)*D*M(L)**T.
+        // and e = [ 1, 1, ..., 1 ]**T.  Note M(A) = M(L)*D*M(L)**T.
         //
-        //        Solve M(L) * x = e.
+        // Solve M(L) * x = e.
         //
         work[1 - 1] = one;
         for (i = 2; i <= n; i = i + 1) {
             work[i - 1] = one + work[(i - 1) - 1] * abs(ef[(i - 1) - 1]);
         }
         //
-        //        Solve D * M(L)**T * x = b.
+        // Solve D * M(L)**T * x = b.
         //
         work[n - 1] = work[n - 1] / df[n - 1];
         for (i = n - 1; i >= 1; i = i - 1) {
             work[i - 1] = work[i - 1] / df[i - 1] + work[(i + 1) - 1] * abs(ef[i - 1]);
         }
         //
-        //        Compute norm(inv(A)) = max(x(i)), 1<=i<=n.
+        // Compute norm(inv(A)) = max(x(i)), 1<=i<=n.
         //
         ix = iRamax(n, work, 1);
         ferr[j - 1] = ferr[j - 1] * abs(work[ix - 1]);
         //
-        //        Normalize error.
+        // Normalize error.
         //
         lstres = zero;
         for (i = 1; i <= n; i = i + 1) {
-            lstres = max(lstres, REAL(abs(x[(i - 1) + (j - 1) * ldx])));
+            lstres = max(lstres, abs(x[(i - 1) + (j - 1) * ldx]));
         }
         if (lstres != zero) {
             ferr[j - 1] = ferr[j - 1] / lstres;
@@ -252,6 +236,6 @@ void Rptrfs(INTEGER const n, INTEGER const nrhs, REAL *d, REAL *e, REAL *df, REA
         //
     }
     //
-    //     End of Rptrfs
+    // End of Rptrfs
     //
 }
