+++ mplapack/reference/Rtprfs.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2008-2021
+ * Copyright (c) 2008-2025
  *      Nakata, Maho
  *      All rights reserved.
  *
@@ -26,6 +26,13 @@
  *
  */
 
+// Derived from LAPACK routine DTPRFS.
+// Original LAPACK authors:
+//   Univ. of Tennessee
+//   Univ. of California Berkeley
+//   Univ. of Colorado Denver
+//   NAG Ltd.
+
 #include <mpblas.h>
 #include <mplapack.h>
 
@@ -51,32 +58,7 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
     INTEGER isave[3];
     REAL lstres = 0.0;
     //
-    //  -- LAPACK computational routine --
-    //  -- LAPACK is a software package provided by Univ. of Tennessee,    --
-    //  -- Univ. of California Berkeley, Univ. of Colorado Denver and NAG Ltd..--
-    //
-    //     .. Scalar Arguments ..
-    //     ..
-    //     .. Array Arguments ..
-    //     ..
-    //
-    //  =====================================================================
-    //
-    //     .. Parameters ..
-    //     ..
-    //     .. Local Scalars ..
-    //     ..
-    //     .. Local Arrays ..
-    //     ..
-    //     .. External Subroutines ..
-    //     ..
-    //     .. Intrinsic Functions ..
-    //     ..
-    //     .. External Functions ..
-    //     ..
-    //     .. Executable Statements ..
-    //
-    //     Test the input parameters.
+    // Test the input parameters.
     //
     info = 0;
     upper = Mlsame(uplo, "U");
@@ -103,7 +85,7 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
         return;
     }
     //
-    //     Quick return if possible
+    // Quick return if possible
     //
     if (n == 0 || nrhs == 0) {
         for (j = 1; j <= nrhs; j = j + 1) {
@@ -119,7 +101,7 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
         transt = 'N';
     }
     //
-    //     NZ = maximum number of nonzero elements in each row of A, plus 1
+    // NZ = maximum number of nonzero elements in each row of A, plus 1
     //
     nz = n + 1;
     eps = Rlamch("Epsilon");
@@ -127,25 +109,25 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
     safe1 = nz * safmin;
     safe2 = safe1 / eps;
     //
-    //     Do for each right hand side
+    // Do for each right hand side
     //
     for (j = 1; j <= nrhs; j = j + 1) {
         //
-        //        Compute residual R = B - op(A) * X,
-        //        where op(A) = A or A**T, depending on TRANS.
+        // Compute residual R = B - op(A) * X,
+        // where op(A) = A or A**T, depending on TRANS.
         //
         Rcopy(n, &x[(j - 1) * ldx], 1, &work[(n + 1) - 1], 1);
         Rtpmv(uplo, trans, diag, n, ap, &work[(n + 1) - 1], 1);
         Raxpy(n, -one, &b[(j - 1) * ldb], 1, &work[(n + 1) - 1], 1);
         //
-        //        Compute componentwise relative backward error from formula
+        // Compute componentwise relative backward error from formula
         //
-        //        max(i) ( abs(R(i)) / ( abs(op(A))*abs(X) + abs(B) )(i) )
+        // max(i) ( abs(R(i)) / ( abs(op(A))*abs(X) + abs(B) )(i) )
         //
-        //        where abs(Z) is the componentwise absolute value of the matrix
-        //        or vector Z.  If the i-th component of the denominator is less
-        //        than SAFE2, then SAFE1 is added to the i-th components of the
-        //        numerator and denominator before dividing.
+        // where abs(Z) is the componentwise absolute value of the matrix
+        // or vector Z.  If the i-th component of the denominator is less
+        // than SAFE2, then SAFE1 is added to the i-th components of the
+        // numerator and denominator before dividing.
         //
         for (i = 1; i <= n; i = i + 1) {
             work[i - 1] = abs(b[(i - 1) + (j - 1) * ldb]);
@@ -153,7 +135,7 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
         //
         if (notran) {
             //
-            //           Compute abs(A)*abs(X) + abs(B).
+            // Compute abs(A)*abs(X) + abs(B).
             //
             if (upper) {
                 kc = 1;
@@ -198,7 +180,7 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
             }
         } else {
             //
-            //           Compute abs(A**T)*abs(X) + abs(B).
+            // Compute abs(A**T)*abs(X) + abs(B).
             //
             if (upper) {
                 kc = 1;
@@ -247,34 +229,34 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
         s = zero;
         for (i = 1; i <= n; i = i + 1) {
             if (work[i - 1] > safe2) {
-                s = max(s, REAL(abs(work[(n + i) - 1]) / work[i - 1]));
+                s = max(s, abs(work[(n + i) - 1]) / work[i - 1]);
             } else {
-                s = max(s, REAL((abs(work[(n + i) - 1]) + safe1) / (work[i - 1] + safe1)));
+                s = max(s, (abs(work[(n + i) - 1]) + safe1) / (work[i - 1] + safe1));
             }
         }
         berr[j - 1] = s;
         //
-        //        Bound error from formula
+        // Bound error from formula
         //
-        //        norm(X - XTRUE) / norm(X) .le. FERR =
-        //        norm( abs(inv(op(A)))*
-        //           ( abs(R) + NZ*EPS*( abs(op(A))*abs(X)+abs(B) ))) / norm(X)
+        // norm(X - XTRUE) / norm(X) .le. FERR =
+        // norm( abs(inv(op(A)))*
+        // ( abs(R) + NZ*EPS*( abs(op(A))*abs(X)+abs(B) ))) / norm(X)
         //
-        //        where
-        //          norm(Z) is the magnitude of the largest component of Z
-        //          inv(op(A)) is the inverse of op(A)
-        //          abs(Z) is the componentwise absolute value of the matrix or
-        //             vector Z
-        //          NZ is the maximum number of nonzeros in any row of A, plus 1
-        //          EPS is machine epsilon
+        // where
+        // norm(Z) is the magnitude of the largest component of Z
+        // inv(op(A)) is the inverse of op(A)
+        // abs(Z) is the componentwise absolute value of the matrix or
+        // vector Z
+        // NZ is the maximum number of nonzeros in any row of A, plus 1
+        // EPS is machine epsilon
         //
-        //        The i-th component of abs(R)+NZ*EPS*(abs(op(A))*abs(X)+abs(B))
-        //        is incremented by SAFE1 if the i-th component of
-        //        abs(op(A))*abs(X) + abs(B) is less than SAFE2.
+        // The i-th component of abs(R)+NZ*EPS*(abs(op(A))*abs(X)+abs(B))
+        // is incremented by SAFE1 if the i-th component of
+        // abs(op(A))*abs(X) + abs(B) is less than SAFE2.
         //
-        //        Use Rlacn2 to estimate the infinity-norm of the matrix
-        //           inv(op(A)) * diag(W),
-        //        where W = abs(R) + NZ*EPS*( abs(op(A))*abs(X)+abs(B) )))
+        // Use Rlacn2 to estimate the infinity-norm of the matrix
+        // inv(op(A)) * diag(W),
+        // where W = abs(R) + NZ*EPS*( abs(op(A))*abs(X)+abs(B) )))
         //
         for (i = 1; i <= n; i = i + 1) {
             if (work[i - 1] > safe2) {
@@ -290,7 +272,7 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
         if (kase != 0) {
             if (kase == 1) {
                 //
-                //              Multiply by diag(W)*inv(op(A)**T).
+                // Multiply by diag(W)*inv(op(A)**T).
                 //
                 Rtpsv(uplo, &transt, diag, n, ap, &work[(n + 1) - 1], 1);
                 for (i = 1; i <= n; i = i + 1) {
@@ -298,7 +280,7 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
                 }
             } else {
                 //
-                //              Multiply by inv(op(A))*diag(W).
+                // Multiply by inv(op(A))*diag(W).
                 //
                 for (i = 1; i <= n; i = i + 1) {
                     work[(n + i) - 1] = work[i - 1] * work[(n + i) - 1];
@@ -308,11 +290,11 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
             goto statement_210;
         }
         //
-        //        Normalize error.
+        // Normalize error.
         //
         lstres = zero;
         for (i = 1; i <= n; i = i + 1) {
-            lstres = max(lstres, REAL(abs(x[(i - 1) + (j - 1) * ldx])));
+            lstres = max(lstres, abs(x[(i - 1) + (j - 1) * ldx]));
         }
         if (lstres != zero) {
             ferr[j - 1] = ferr[j - 1] / lstres;
@@ -320,6 +302,6 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
         //
     }
     //
-    //     End of Rtprfs
+    // End of Rtprfs
     //
 }
